requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-filter==23.5
django-jazzmin==3.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.10
whitenoise==6.5.0
dj-database-url==1.0.0

# Test tools
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
